% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog}
\alias{catalog}
\title{Create a data source catalog}
\usage{
catalog(source, engine, pattern = NULL, filter = NULL, import_specs = NULL)
}
\arguments{
\item{source}{The source for the data.  This parameter is required.}

\item{engine}{The data engine to use for this data source. This parameter
is required.}

\item{pattern}{A pattern to use when loading the data source.  The pattern
can be a name or a vector of names.  Names also accept wildcards.}

\item{filter}{One or more filter expressions to use when fetching 
the data. This filter will apply to all fetch operations on this catalog.}

\item{import_specs}{The import specs to use for any fetch operation on 
this catalog.}
}
\value{
The loaded data catalog.
}
\description{
The \code{catalog} function returns a data catalog
for a data source.
}
\examples{
# Create temp directory
tmp <- tempdir()

}
\seealso{
The \code{\link{fetch}} function to retrieve data from the catalog.
}
