% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\docType{data}
\encoding{UTF-8}
\name{engines}
\alias{engines}
\title{A list of engine types}
\format{
An object of class \code{etype} of length 9.
}
\usage{
engines
}
\description{
The engines enumeration contains all possible options
for the "engine" parameter of the \code{\link{catalog}} function.  Use
this enumeration to specify what kind of data you would like to load.
Options are: csv, dbf, rda, rds, rdata, sas7bdat, xls, xlsx, and xpt.
}
\keyword{datasets}
