% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fetch.R
\name{fetch}
\alias{fetch}
\title{Synonym function of read.data}
\usage{
fetch(
  file,
  type,
  skip = 0,
  range = NULL,
  dec = ".",
  sep,
  sheet = NULL,
  col_types = NULL,
  pattern_filename,
  return = T
)
}
\description{
Synonym function of read.data
}
