% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\docType{package}
\encoding{UTF-8}
\name{fetchpackage}
\alias{fetchpackage}
\title{Fetch data from many data sources}
\description{
The \strong{fetch} package allows you to retrieve data from many different
data sources.  The package retrieves data in a memory-efficient manner.
You first identify the data by defining a data catalog.  Then fetch
the data from the catalog.  Catalogs can be defined for many popular 
data formats: csv, rds, sas7bdat, excel, etc.

The functions contained in the \strong{fetch} package are as follows:
\itemize{
  \item{\code{\link{catalog}}: Creates a data library}
  \item{\code{\link{fetch}}: Creates a data dictionary}
  \item{\code{\link{import_spec}}: Defines an import spec for a specific dataset}
}
}
\keyword{internal}
