% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_test.R
\name{global_test}
\alias{global_test}
\title{Fetches global test statistic}
\usage{
global_test(matrix_samples, conditions)
}
\arguments{
\item{conditions}{A vector of conditions}

\item{samples_used}{An array of CLR samples}
}
\value{
A single test statistic and p-value.
}
\description{
Tabulates the global test statistic according to Hoetelling's T2
}
