% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{print.dcat}
\alias{print.dcat}
\title{Print a data catalog}
\usage{
\method{print}{dcat}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The library to print.}

\item{...}{Any follow-on parameters.}

\item{verbose}{Whether or not to print the accessor in verbose style.
By default, the parameter is FALSE, meaning to print in summary style.}
}
\value{
The object, invisibly.
}
\description{
A class-specific instance of the \code{print} function for 
data accessors.  The function prints the accessor in a summary manner.  
Use \code{verbose = TRUE} to print the accessor as a list.
}
\examples{
# Create temp directory
tmp <- tempdir()

}
