% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinfo.R
\name{print.dinfo}
\alias{print.dinfo}
\title{Print a data catalog item}
\usage{
\method{print}{dinfo}(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{The library to print.}

\item{...}{Any follow-on parameters.}

\item{verbose}{Whether or not to print the info in verbose style.
By default, the parameter is FALSE, meaning to print in summary style.
Verbose style includes a full data dictionary and printing of all 
attributes.}
}
\value{
The data catalog object, invisibly.
}
\description{
A class-specific instance of the \code{print} function for 
data catalog items.  The function prints the info in a summary manner.  
Use \code{verbose = TRUE} to print the data info as a list.
}
