% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - fetch.R
\name{read.data}
\alias{read.data}
\title{function for simplified reading of files}
\usage{
read.data(
  file,
  type,
  skip = 0,
  range = NULL,
  dec = ".",
  sep,
  sheet = NULL,
  col_types = NULL
)
}
\arguments{
\item{file}{Filename or URL for google sheet}

\item{type}{File type. Usually not necessary since this is checked automatically.}

\item{skip}{How many rows to skip in the file}

\item{range}{Only for google sheets}

\item{dec}{decimal separator}

\item{sep}{value separator}

\item{sheet}{for google sheets and excel sheets. Specify which sheet in the file to read.}

\item{col_types}{for google sheets. Specify variable types for columns. See documentation for read_sheet() in googlesheets4}
}
\description{
Tries to automaticaly handle .txt, .csv, .xls, .xlsx and google sheet links
}
