% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\docType{package}
\name{fetch}
\alias{fetch}
\title{Fetch data from many data sources}
\usage{
fetch(catalog, filter = NULL, top = NULL, import_specs = NULL)
}
\arguments{
\item{catalog}{The catalog item to fetch data from.}

\item{filter}{An expression to be used to filter the fetched data.}

\item{top}{A number of records to return from the data source.  Valid 
value is an integer.}

\item{import_specs}{The import specs to use for the fetch operation. Import
specs can be used to control the data types.}
}
\description{
The \strong{fetch} package allows you to retrieve data from many different
data sources.  The package retrieves data in a memory-efficient manner.
You first identify the data by defining a data accessor.  Then fetch
the data from the accessor.  Accessors can be defined for many popular 
data formats: csv, rds, sas7bdat, excel, etc.

The functions contained in the \strong{fetch} package are as follows:
\itemize{
  \item{\code{\link{catalog}}: Creates a data library}
  \item{\code{\link{fetch}}: Creates a data dictionary}
  \item{\code{\link{import_spec}}: Defines an import spec for a specific dataset}
}
}
